Updated GetTime to return a double for more precision. + Updated Application::Statistics to have better values and for Application::Run to use the double instead now.

+++ include/utility/platform_utils.h
@@ -5,7 +5,7 @@ namespace penumbra
 	class CPlatformUtils
 	{
 	public:
-		static float GetTime();
+		static double GetTime();
 
 		static std::filesystem::path OpenFileDialog(const char* filter);
 		static std::filesystem::path SaveFileDialog(const char* filter);
